Split the Swarm from the NetworkBehaviour
